backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Security - Rate Limiting
slowapi==0.1.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
psycopg2-binary==2.9.9

# Temporal
temporalio==1.4.0

# OpenAI
openai==1.12.0

# Environment and settings
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS
python-jose==3.3.0

# Utilities
httpx==0.26.0
python-dateutil==2.8.2

# PERFORMANCE: Redis for caching and distributed rate limiting
redis==5.0.1
limits[async,redis]==3.6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1

